feat: migrate sources.js to typescript
