+++ src/components/view/sources/sources.ts
@@ -1,0 +1,22 @@
+import './sources.css';
+import {Source} from '../../../types/types'
+
+class Sources {
+    draw(data: Source[]) {
+        const fragment = document.createDocumentFragment();
+        const sourceItemTemp = document.querySelector('#sourceItemTemp') as HTMLTemplateElement;
+
+        data.forEach((item) => {
+            const sourceClone = sourceItemTemp.content.cloneNode(true) as Element;
+
+            sourceClone.querySelector('.source__item-name')!.textContent = item.name;
+            sourceClone.querySelector('.source__item')!.setAttribute('data-source-id', item.id);
+
+            fragment.append(sourceClone);
+        });
+
+        document.querySelector('.sources')!.append(fragment);
+    }
+}
+
+export default Sources;
